requirements.txt
argon2-cffi
colored
cryptography
